refactor(InputView): 자동차 이름과 전진 횟수 입력받는 클래스 분리
